+++ openrouteservice/src/main/java/heigit/ors/services/routing/requestprocessors/gpx/GpxRoutingResponseWriter.java
@@ -14,18 +14,18 @@ import heigit.ors.routing.WeightingMethod;
 import heigit.ors.services.routing.RoutingServiceSettings;
 import heigit.ors.util.AppInfo;
 import heigit.ors.util.GeomUtility;
-import heigit.ors.util.gpxUtil.BoundsType;
-import heigit.ors.util.gpxUtil.CopyrightType;
-import heigit.ors.util.gpxUtil.EmailType;
-import heigit.ors.util.gpxUtil.Gpx;
-import heigit.ors.util.gpxUtil.GpxExtensions;
-import heigit.ors.util.gpxUtil.LinkType;
-import heigit.ors.util.gpxUtil.MetadataType;
-import heigit.ors.util.gpxUtil.PersonType;
-import heigit.ors.util.gpxUtil.RteType;
-import heigit.ors.util.gpxUtil.RteTypeExtensions;
-import heigit.ors.util.gpxUtil.WptType;
-import heigit.ors.util.gpxUtil.WptTypeExtensions;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.BoundsType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.CopyrightType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.EmailType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.Gpx;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.GpxExtensions;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.LinkType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.MetadataType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.PersonType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.RteType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.RteTypeExtensions;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.WptType;
+import heigit.ors.util.GlobalExportWriter.gpxUtil.beans.WptTypeExtensions;
 
 import javax.xml.datatype.DatatypeFactory;
 import javax.xml.datatype.XMLGregorianCalendar;
