+++ openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/GlobalExportProcessor.java
@@ -1,0 +1,9 @@
+package heigit.ors.util.GlobalExportWriter;
+
+import heigit.ors.routing.RoutingRequest;
+
+public class GlobalExportProcessor {
+
+    public GlobalExportProcessor(RoutingRequest routingRequest) {
+    }
+}
