+++ openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/gpxUtil/beans/GpxExtensions.java
@@ -1,4 +1,4 @@
-package heigit.ors.util.gpxUtil;
+package heigit.ors.util.GlobalExportWriter.gpxUtil.beans;
 
 
 import javax.xml.bind.annotation.XmlAccessType;
