+++ openrouteservice/src/main/java/heigit/ors/util/GlobalExportWriter/gpxUtil/beans/PtType.java
@@ -1,5 +1,5 @@
 
-package heigit.ors.util.gpxUtil;
+package heigit.ors.util.GlobalExportWriter.gpxUtil.beans;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
